+++ demo/md-enhance/src/.vuepress/config.ts
@@ -36,11 +36,14 @@ export default defineUserConfig({
         "/demo/mark",
         "/demo/include",
         "/demo/mermaid",
+        "/demo/playground",
         "/demo/presentation",
+        "/demo/stylized",
         "/demo/sup-sub",
         "/demo/tabs",
         "/demo/tasklist",
         "/demo/tex",
+        "/demo/vue-playground",
       ],
     },
 
@@ -56,10 +59,45 @@ export default defineUserConfig({
 
   plugins: [
     mdEnhancePlugin({
-      enableAll: true,
+      align: true,
+      attrs: true,
+      chart: true,
+      codetabs: true,
+      container: true,
+      demo: true,
+      echarts: true,
+      flowchart: true,
+      gfm: true,
+      imageSize: true,
+      include: true,
+      lazyLoad: true,
+      mark: true,
+      mermaid: true,
+      playground: {
+        presets: ["ts", "vue"],
+      },
       presentation: {
         plugins: ["highlight", "math", "search", "notes", "zoom"],
       },
+      stylize: [
+        {
+          matcher: "Recommanded",
+          replacer: ({ tag }) => {
+            if (tag === "em")
+              return {
+                tag: "Badge",
+                attrs: { type: "tip" },
+                content: "Recommanded",
+              };
+          },
+        },
+      ],
+      sub: true,
+      sup: true,
+      tabs: true,
+      tex: true,
+      vpre: true,
+      vuePlayground: true,
     }),
   ],
 });