+++ packages/comment2/src/client/components/Waline.ts
@@ -1,6 +1,6 @@
 import { usePageFrontmatter, usePageLang, withBase } from "@vuepress/client";
-import { Waline } from "@waline/client/dist/component";
-import { pageviewCount } from "@waline/client/dist/pageview";
+import { Waline } from "@waline/client/dist/component.mjs";
+import { pageviewCount } from "@waline/client/dist/pageview.mjs";
 import { computed, defineComponent, h, onMounted, watch } from "vue";
 import { useRoute } from "vue-router";
 import { useLocaleConfig } from "vuepress-shared/lib/client";
@@ -19,11 +19,12 @@ declare const COMMENT_OPTIONS: WalineOptions;
 
 declare const WALINE_LOCALES: WalineLocaleConfig;
 
-export const walineOption = COMMENT_OPTIONS;
+const walineOption = COMMENT_OPTIONS;
+const walineLocales = WALINE_LOCALES;
+const enableWaline = Boolean(walineOption.serverURL);
 
-export const enableWaline = Boolean(walineOption.serverURL);
-
-export const walineLocales = WALINE_LOCALES;
+if (walineOption.metaIcon !== false)
+  import("@waline/client/dist/waline-meta.css");
 
 export default defineComponent({
   name: "WalineComment",