+++ packages/create/template/i18n/.vuepress/theme.ts
@@ -1,6 +1,6 @@
 import { hopeTheme } from "vuepress-theme-hope";
-import * as navbar from "./navbar.js";
-import * as sidebar from "./sidebar.js";
+import * as navbar from "./navbar/index.js";
+import * as sidebar from "./sidebar/index.js";
 
 export default hopeTheme({
   hostname: "https://vuepress-theme-hope-v2-demo.mrhope.site",
@@ -132,10 +132,45 @@ export default hopeTheme({
     },
 
     mdEnhance: {
-      enableAll: true,
+      align: true,
+      attrs: true,
+      chart: true,
+      codetabs: true,
+      container: true,
+      demo: true,
+      echarts: true,
+      flowchart: true,
+      gfm: true,
+      imageSize: true,
+      include: true,
+      lazyLoad: true,
+      mark: true,
+      mermaid: true,
+      playground: {
+        presets: ["ts", "vue"],
+      },
       presentation: {
         plugins: ["highlight", "math", "search", "notes", "zoom"],
       },
+      stylize: [
+        {
+          matcher: "Recommanded",
+          replacer: ({ tag }) => {
+            if (tag === "em")
+              return {
+                tag: "Badge",
+                attrs: { type: "tip" },
+                content: "Recommanded",
+              };
+          },
+        },
+      ],
+      sub: true,
+      sup: true,
+      tabs: true,
+      tex: true,
+      vpre: true,
+      vuePlayground: true,
     },
   },
 });